Revise few types and interfaces to avoid implicit integral type conversions

+++ src/thorin/type.h
@@ -45,7 +45,7 @@ private:
 /// Type variable.
 class Var : public Type {
 private:
-    Var(TypeTable& table, int depth)
+    Var(TypeTable& table, size_t depth)
         : Type(table, Node_Var, {})
         , depth_(depth)
     {
@@ -53,7 +53,7 @@ private:
     }
 
 public:
-    int depth() const { return depth_; }
+    size_t depth() const { return depth_; }
     virtual std::ostream& stream(std::ostream&) const override;
 
 private:
@@ -62,7 +62,7 @@ private:
     virtual const Type* vrebuild(TypeTable& to, Types ops) const override;
     virtual const Type* vreduce(int, const Type*, Type2Type&) const override;
 
-    int depth_;
+    size_t depth_;
 
     friend class TypeTable;
 };
@@ -365,7 +365,7 @@ class TypeTable : public TypeTableBase<Type> {
 public:
     TypeTable();
 
-    const Var* var(int depth) { return unify(new Var(*this, depth)); }
+    const Var* var(size_t depth) { return unify(new Var(*this, depth)); }
     const Lambda* lambda(const Type* body, const char* name) { return unify(new Lambda(*this, body, name)); }
     const Type* app(const Type* callee, const Type* arg);
